+++ src/pages/profile/appointments/Appointments.tsx
@@ -1,0 +1,14 @@
+import { Box, styled } from '@mui/material';
+const MainContainer = styled(Box)(() => ({
+    width: "50%",
+  }));
+
+function Appointments() {
+  return (
+    <MainContainer>
+        appointments
+    </MainContainer>
+  )
+}
+
+export default Appointments
